refactor: action클래스 이동
